backend/requirements.txt
Django==4.2.27
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.1.0
PyJWT
psycopg2-binary
gunicorn
dj-database-url
whitenoise
python-dotenv
vanna
openai
chromadb
